+++ semantic_navigation_bt/src/action/generate_random_goals_service.cpp
@@ -40,7 +40,7 @@ BT::NodeStatus GenerateRandomGoalsService::on_completion(
   std::shared_ptr<semantic_navigation_msgs::srv::GenerateRandomGoals::Response> response)
 {
   BT::NodeStatus status = BT::NodeStatus::FAILURE;
-  if (response->goals.poses.size() > 0) {
+  if (response->goals.size() > 0) {
     setOutput("goals", response->goals);
     status = BT::NodeStatus::SUCCESS;
   }