+++ semantic_navigation_bt/test/action/test_generate_random_goals_service.cpp
@@ -43,7 +43,7 @@ public:
   {
     (void)request_header;
     (void)request;
-    response->goals.poses.push_back(geometry_msgs::msg::Pose());
+    response->goals.push_back(geometry_msgs::msg::PoseStamped());
   }
 };
 
@@ -60,18 +60,13 @@ public:
     // Create the blackboard that will be shared by all of the nodes in the tree
     config_->blackboard = BT::Blackboard::create();
     // Put items on the blackboard
-    config_->blackboard->set(
-      "node",
-      node_);
+    config_->blackboard->set("node", node_);
     config_->blackboard->set<std::chrono::milliseconds>(
-      "server_timeout",
-      std::chrono::milliseconds(20));
+      "server_timeout", std::chrono::milliseconds(20));
     config_->blackboard->set<std::chrono::milliseconds>(
-      "bt_loop_duration",
-      std::chrono::milliseconds(10));
+      "bt_loop_duration", std::chrono::milliseconds(10));
     config_->blackboard->set<std::chrono::milliseconds>(
-      "wait_for_service_timeout",
-      std::chrono::milliseconds(1000));
+      "wait_for_service_timeout", std::chrono::milliseconds(1000));
 
     factory_->registerNodeType<semantic_navigation_bt::GenerateRandomGoalsService>(
       "GenerateRandomGoals");
@@ -131,8 +126,8 @@ TEST_F(GenerateRandomGoalsServiceTestFixture, test_tick)
   EXPECT_EQ(tree_->rootNode()->executeTick(), BT::NodeStatus::SUCCESS);
 
   // Check if the output is correct
-  auto goals = config_->blackboard->get<geometry_msgs::msg::PoseArray>("goals");
-  EXPECT_EQ(goals.poses.size(), 1);
+  auto goals = config_->blackboard->get<std::vector<geometry_msgs::msg::PoseStamped>>("goals");
+  EXPECT_EQ(goals.size(), 1);
 }
 
 int main(int argc, char ** argv)