+++ semantic_navigation_tasks/include/semantic_navigation_tasks/task_services.hpp
@@ -29,6 +29,7 @@
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_lifecycle/lifecycle_publisher.hpp"
 #include "nav2_util/lifecycle_node.hpp"
+#include "geometry_msgs/msg/pose_array.hpp"
 #include "geometry_msgs/msg/pose_stamped.hpp"
 #include "nav_msgs/msg/occupancy_grid.hpp"
 #include "polygon_msgs/msg/polygon2_d_collection.hpp"
@@ -163,9 +164,9 @@ protected:
    * @param n Number of goals.
    * @param region Region of interest.
    * @param limits Limits of the cells.
-   * @return geometry_msgs::msg::PoseArray Goals.
+   * @return std::vector<geometry_msgs::msg::PoseStamped> Goals.
    */
-  geometry_msgs::msg::PoseArray generateRandomGoals(
+  std::vector<geometry_msgs::msg::PoseStamped> generateRandomGoals(
     unsigned int n, Region region, CellLimits limits);
 
   /**